Fix: Hyperlinks break when removing rows
When removing a row from above cells with hyperlinks, HyperlinkCollection gets broken due to the reversed order of operation. The reversing of array (in current code) works as expected only when inserting rows ($pNumRows/Cols is positive), but breaks when removing rows ($pNumRows/Cols is negative). Therefore, reversing should be conditional. Not sure if this is the best solution, but it works. The same should most likely apply for other reference alteration codes in this file where array_reverse (or reverse order of operation) is used.