Drop official support of PHP 5.2
Since unit tests are broken with PHP 5.2 and that we can barely keep up with
unit tests for recent version of PHP 5.3+, we drop PHP 5.2 tests. And thus
we cannot advertised PHPExcel as working on 5.2.

Moreover PHP 5.2 is likely to be broken since the introduction of timezone more
than one year ago, in commit 53a32ce7.

However, PHP 5.2 specific code was not dropped, hence PHPExcel will actually keep
working as much as it did until now.