Flush the XMLWriter every 100 XML elements.
The PHP XMLWriter class doesn't automatically flush out the generated
XML document into the associated files. As the code doesn't explicitely
call the "flush()" method anywhere, it means the document is only
flushed once, when closing the document.

To make disk caching effective in reducing memory usage, this commit
overrides some methods of XMLWriter in order to trigger an automatic
flush every 100 XML elements, in order to keep the memory used by the
XMLWriter as low as possible.