Clarify prerequisites
Clean up extension names to match the actual PHP extension names (eg. 'zip' rather than 'php_zip'), trying to be distribution-agnostic.

Add note that we need all three of xml, xmlreader and xmlwriter - on many distributions these are a single package to install, but on FreeBSD these are distinct packages (as they indeed are when compiling PHP from source.)

Change note about GD needing version 2.  Version 2 has been the default version to compile the gd extension against for many years, but saying we need an extension 'gd2' misleads new users to thinking that there is a distinct extension to install.