Changed disk cache default dir to system temp dir
Previously if you turned on disk caching and didn't specify a directory current working directory was used "./" (property default value). Which is not really a desired or expected behaviour as this value is optional. If left unset phpexcel tries to write temporary xml files to CWD of your application and fails if it is not writeable, which it should be in normal setup or creates many temporary xml files in your root.