Suppress error for integer data-series
In my development environment, line 301 throws a warning saying: "Cannot use a scalar value as an array", but the operation completes successfully. I believe that is because the second check on that line assumes that the value is a string, which it isn't in the case of an integer. Adding a is_string check fixes the error.