ComputerTinker Patch Rollup 2016-06-17
* Allow comma separated chart data series values using the format:
'Worksheet Name'!$C$10,'Worksheet Name'!$C$13

* Add more color definitions to the pre-defined constant values,
limiting the expansion to combinations of the FF and 80 hex codes.

* Numerous fixes for HTML rendering: Don't output extra TR and TD at the
bottom unless either a chart or a drawing were found. Style hyperlinks
using same style info as the TD, as intended. Write unified border css
property if all of the border-* components are identical. When TD
specifies border:none, do not write out that property as it overrides
the table gridlines. When no cell background fill is specified, do not
write out a background color of white since the page background has
already been set to white. Fix a typo where the TDs were trying to
reference a non-existent 'column' class instead of the 'col' class which
is actually defined.

* Add support for proper output of boolean values and numeric values
treated as
logicals.